Move theme tracker in to its own file and namespace.
